created style-utils helper collection
