+++ src/app/utils/utils.ts
@@ -123,12 +123,6 @@ export function parseJwt(token): {
   return JSON.parse(jsonPayload);
 }
 
-export function addAlphaToHexColor(color, opacity) {
-  // coerce values so it is between 0 and 1.
-  const opacity1 = Math.round(Math.min(Math.max(opacity || 1, 0), 1) * 255);
-  return color + opacity1.toString(16).toUpperCase();
-}
-
 /**
  * This is a simple shorthand function to create factories for services.
  * The use case is, when multiple services extends the same class and one of these services will be provided.
